通信完善，考虑使用websocket

+++ src/com/lh/dao/dao_sysuser.java
@@ -3,5 +3,5 @@ package com.lh.dao;
 
 import com.lh.vo.sysuser;
 
-public interface dao_sysuser extends baseDao<sysuser, Integer> {
+public interface dao_sysuser extends baseDao<sysuser, Long> {
 }
