+++ src/com/lh/daoImp/sysuserDaoImp.java
@@ -3,6 +3,6 @@ package com.lh.daoImp;
 import com.lh.dao.dao_sysuser;
 import com.lh.vo.sysuser;
 
-public class sysuserDaoImp extends baseDaoImp<sysuser, Integer> implements dao_sysuser{
+public class sysuserDaoImp extends baseDaoImp<sysuser, Long> implements dao_sysuser{
     
 }
